Getting to a functional test.
Will need a significant amount of rearranging to get testable code.
Currently too much control is assumed by the irc.connect() function to
allow for simple testing.  Either this must be sorted out, or a more
predictable way of listening for certain events in order must be planned
for.

Currently it works fine, but a setTimeout is needed to force the tests
to quit after a time.